refactor(concealer): stop using utility directory
