+++ external/concealer/include/concealer/field_operator_application_for_autoware_universe.hpp
@@ -26,9 +26,9 @@
 #include <concealer/autoware_universe.hpp>
 #include <concealer/field_operator_application.hpp>
 #include <concealer/task_queue.hpp>
-#include <concealer/utility/publisher_wrapper.hpp>
-#include <concealer/utility/service_with_validation.hpp>
-#include <concealer/utility/subscriber_wrapper.hpp>
+#include <concealer/publisher_wrapper.hpp>
+#include <concealer/service_with_validation.hpp>
+#include <concealer/subscriber_wrapper.hpp>
 #include <geometry_msgs/msg/pose_with_covariance_stamped.hpp>
 #include <tier4_external_api_msgs/msg/emergency.hpp>
 #include <tier4_external_api_msgs/srv/engage.hpp>
