+++ external/concealer/include/concealer/subscriber_wrapper.hpp
