+++ external/concealer/include/concealer/visibility.hpp
@@ -12,8 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef CONCEALER__UTILITY__VISIBILITY_HPP_
-#define CONCEALER__UTILITY__VISIBILITY_HPP_
+#ifndef CONCEALER__VISIBILITY_HPP_
+#define CONCEALER__VISIBILITY_HPP_
 
 #ifdef __cplusplus
 extern "C" {
@@ -57,4 +57,4 @@ extern "C" {
 }
 #endif
 
-#endif  // CONCEALER__UTILITY__VISIBILITY_HPP_
+#endif  // CONCEALER__VISIBILITY_HPP_
